refactor: move table cell types to composition package for reuse with data_table
