+++ slack-api-client/src/main/java/com/slack/api/util/json/GsonFactory.java
@@ -11,7 +11,7 @@ import com.slack.api.model.admin.AppWorkflow;
 import com.slack.api.model.block.ContextBlockElement;
 import com.slack.api.model.block.ContextActionsBlockElement;
 import com.slack.api.model.block.LayoutBlock;
-import com.slack.api.model.block.TableCell;
+import com.slack.api.model.block.composition.TableCell;
 import com.slack.api.model.block.composition.TextObject;
 import com.slack.api.model.block.element.BlockElement;
 import com.slack.api.model.block.element.RichTextElement;
