+++ slack-api-model/src/main/java/com/slack/api/model/block/RichTextBlock.java
@@ -16,7 +16,7 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RichTextBlock implements LayoutBlock, TableCell {
+public class RichTextBlock implements LayoutBlock, com.slack.api.model.block.composition.TableCell {
     public static final String TYPE = "rich_text";
     private final String type = TYPE;
     @Builder.Default
