+++ slack-api-model/src/main/java/com/slack/api/model/block/TableBlock.java
@@ -5,17 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import com.slack.api.model.block.composition.TableCell;
+
 import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Displays tabular data as a grid of rows and cells.
- *
- * <p>Each row is a list of {@link TableCell cells}; a cell may be a
- * {@link RawTextTableCell} ({@code raw_text}), {@link RawNumberTableCell}
- * ({@code raw_number}), or a {@link RichTextBlock} ({@code rich_text}). A table
- * supports up to 100 rows and up to 20 cells per row, with optional per-column
- * behavior configured via {@link TableColumnSetting column settings}.</p>
  *
  * @see <a href="https://docs.slack.dev/reference/block-kit/blocks/table-block">Table block</a>
  */
