+++ slack-api-model/src/test/java/test_locally/unit/GsonFactory.java
@@ -8,7 +8,7 @@ import com.slack.api.model.File;
 import com.slack.api.model.block.ContextBlockElement;
 import com.slack.api.model.block.ContextActionsBlockElement;
 import com.slack.api.model.block.LayoutBlock;
-import com.slack.api.model.block.TableCell;
+import com.slack.api.model.block.composition.TableCell;
 import com.slack.api.model.block.composition.TextObject;
 import com.slack.api.model.block.element.BlockElement;
 import com.slack.api.model.block.element.RichTextElement;
